requirements.txt
pandas
numpy
matplotlib
plotly
seaborn
nbformat
chainlit
python-dotenv
protobuf
google-generativeai
langchain
langchain_core
langchain_experimental
langchain_google_genai
langchain_community
langchain_openai
langchain_community
langserve
langchain-groq
langchainhub
fastapi
uvicorn
sse_starlette
bs4
pypdf
PyPDF2
chromadb
faiss-cpu
pinecone-client
